CMake change for build using instaled 3rd party libraries
