--- tmp_stash/appveyor.yml
@@ -1,38 +1,0 @@
-version: 1.0.{build}
-image: Visual Studio 2017
-init:
-- cmd: 
-environment:
-  matrix:
-  - BITS: 64
-  - BITS: 32
-install:
-- cmd: >-
-    call "C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Auxiliary\Build\vcvars%BITS%.bat"
-
-    mkdir %APPVEYOR_BUILD_FOLDER%\win\w%BITS%
-
-    appveyor-retry appveyor DownloadFile https://github.com/KxSystems/kdb/raw/master/w%BITS%/q.lib -FileName %APPVEYOR_BUILD_FOLDER%\win\w%BITS%\q.lib
-build_script:
-- cmd: nmake /f win\Makefile all%BITS%
-after_build:
-- cmd: mkdir ffi
-- ps: Copy-Item w$env:BITS -Destination .\ffi\w$env:BITS -Recurse
-- ps: Copy-Item ffi.q -Destination .\ffi
-- ps: Copy-Item test_win.q -Destination .\ffi
-- ps: Copy-Item LICENSE -Destination .\ffi\LICENSE.ffi
-- ps: Copy-Item README.md -Destination .\ffi\README.ffi
-- IF /I "%APPVEYOR_REPO_TAG%" == "false" set APPVEYOR_REPO_TAG_NAME=%APPVEYOR_REPO_BRANCH%
-- cmd: 7z a ffi_win%BITS%-%APPVEYOR_REPO_TAG_NAME%.zip %APPVEYOR_BUILD_FOLDER%\ffi
-artifacts:
-- path: 'ffi_win*.zip'
-
-deploy:
-  provider: GitHub
-  draft: true
-  auth_token:
-    secure: P2U9jL8L7es2Iv/SSthmQ+a9qSZ41OSiGHjSdOMve7eKwWoo3zh/8DvJqkkMfhS7
-  prerelease: true
-  on:
-    branch: master                 # release from master branch only
-    appveyor_repo_tag: true        # deploy on tag push only
