--- tmp_stash/win/vcpkg/info/dlfcn-win32_1.1.1-1_x64-windows-static.list
@@ -1,15 +1,0 @@
-x64-windows-static/
-x64-windows-static/debug/
-x64-windows-static/debug/lib/
-x64-windows-static/debug/lib/dl.lib
-x64-windows-static/include/
-x64-windows-static/include/dlfcn.h
-x64-windows-static/lib/
-x64-windows-static/lib/dl.lib
-x64-windows-static/share/
-x64-windows-static/share/dlfcn-win32/
-x64-windows-static/share/dlfcn-win32/copyright
-x64-windows-static/share/dlfcn-win32/dlfcn-win32-config.cmake
-x64-windows-static/share/dlfcn-win32/dlfcn-win32-targets-debug.cmake
-x64-windows-static/share/dlfcn-win32/dlfcn-win32-targets-release.cmake
-x64-windows-static/share/dlfcn-win32/dlfcn-win32-targets.cmake
