--- tmp_stash/win/vcpkg/info/dlfcn-win32_1.1.1-1_x86-windows-static.list
@@ -1,15 +1,0 @@
-x86-windows-static/
-x86-windows-static/debug/
-x86-windows-static/debug/lib/
-x86-windows-static/debug/lib/dl.lib
-x86-windows-static/include/
-x86-windows-static/include/dlfcn.h
-x86-windows-static/lib/
-x86-windows-static/lib/dl.lib
-x86-windows-static/share/
-x86-windows-static/share/dlfcn-win32/
-x86-windows-static/share/dlfcn-win32/copyright
-x86-windows-static/share/dlfcn-win32/dlfcn-win32-config.cmake
-x86-windows-static/share/dlfcn-win32/dlfcn-win32-targets-debug.cmake
-x86-windows-static/share/dlfcn-win32/dlfcn-win32-targets-release.cmake
-x86-windows-static/share/dlfcn-win32/dlfcn-win32-targets.cmake
