--- tmp_stash/win/vcpkg/info/libffi_3.1-1_x64-windows-static.list
@@ -1,12 +1,0 @@
-x64-windows-static/
-x64-windows-static/debug/
-x64-windows-static/debug/lib/
-x64-windows-static/debug/lib/libffi.lib
-x64-windows-static/include/
-x64-windows-static/include/ffi.h
-x64-windows-static/include/ffitarget.h
-x64-windows-static/lib/
-x64-windows-static/lib/libffi.lib
-x64-windows-static/share/
-x64-windows-static/share/libffi/
-x64-windows-static/share/libffi/copyright
