+++ travis_win/libffi/install/bin/libffi.def
@@ -1,0 +1,39 @@
+EXPORTS 
+ffi_call 
+ffi_call_go 
+ffi_closure_alloc 
+ffi_closure_free 
+ffi_get_struct_offsets 
+ffi_java_ptrarray_to_raw 
+ffi_java_raw_call 
+ffi_java_raw_size 
+ffi_java_raw_to_ptrarray 
+ffi_prep_cif 
+ffi_prep_cif_var 
+ffi_prep_closure 
+ffi_prep_closure_loc 
+ffi_prep_go_closure 
+ffi_prep_java_raw_closure 
+ffi_prep_java_raw_closure_loc 
+ffi_prep_raw_closure 
+ffi_prep_raw_closure_loc 
+ffi_ptrarray_to_raw 
+ffi_raw_call 
+ffi_raw_size 
+ffi_raw_to_ptrarray 
+ffi_type_complex_double 
+ffi_type_complex_float 
+ffi_type_complex_longdouble 
+ffi_type_double 
+ffi_type_float 
+ffi_type_longdouble 
+ffi_type_pointer 
+ffi_type_sint16 
+ffi_type_sint32 
+ffi_type_sint64 
+ffi_type_sint8 
+ffi_type_uint16 
+ffi_type_uint32 
+ffi_type_uint64 
+ffi_type_uint8 
+ffi_type_void
