+++ travis_win/libffi/install/lib/pkgconfig/libffi.pc
@@ -1,0 +1,11 @@
+prefix=/home/mshimizu/repos/libffi/build/../install
+exec_prefix=${prefix}
+libdir=${exec_prefix}/lib
+toolexeclibdir=${libdir}/../lib
+includedir=${prefix}/include
+
+Name: libffi
+Description: Library supporting Foreign Function Interfaces
+Version: 3.3
+Libs: -L${toolexeclibdir} -lffi
+Cflags: -I${includedir}
